FAD-6207 finished code, time to write tests
