+++ src/config/routes.js
@@ -381,11 +381,6 @@ const routes = [
     condition: configFlag('featureFlags.has_signup')
   },
   {
-    path: '/onboarding/dedicated-ips',
-    component: onboarding.DedicatedIps,
-    condition: configFlag('featureFlags.has_signup')
-  },
-  {
     path: '/onboarding/sending-domain',
     component: onboarding.SendingDomainPage,
     condition: configFlag('featureFlags.has_signup')
