+++ src/pages/billing/ChangePlanPage.js
@@ -17,6 +17,25 @@ export class ChangePlanPage extends Component {
     this.props.getBillingCountries();
   }
 
+  onSubmit = (values, useSavedCC) => {
+    const { account, updateSubscription, billingCreate, billingUpdate, showAlert, history } = this.props;
+
+    // decides which action to be taken based on
+    // if it already has billing and if you use a saved CC
+    const action = account.billing
+      ? (
+        useSavedCC
+          ? updateSubscription(values.planpicker.code)
+          : billingUpdate(values))
+      : billingCreate(values); // creates Zuora account
+
+
+    return action
+      .then(() => history.push('/account/billing'))
+      .then(() => showAlert({ type: 'success', message: 'Subscription Updated' }))
+      .catch((err) => showAlert({ type: 'error', message: 'Plan Update Failed', details: err.message }));
+  }
+
   render() {
 
     if (this.props.loading) {
@@ -27,7 +46,7 @@ export class ChangePlanPage extends Component {
       <Page breadcrumbAction={{ content: 'Back to billing', to: '/account/billing', Component: Link }}>
         <PendingPlanBanner account={this.props.account} />
         <SuspendedBanner account={this.props.account}/>
-        { this.props.canChangePlan && <ChangePlan /> }
+        { this.props.canChangePlan && <ChangePlan onSubmit={this.onSubmit} /> }
       </Page>
     );
   }
