+++ src/pages/onboarding/ChoosePlan.js
@@ -1,17 +1,35 @@
 import React, { Component, Fragment } from 'react';
 import { connect } from 'react-redux';
 import { Panel } from '@sparkpost/matchbox';
+import { showAlert } from 'src/actions/globalAlert';
 import { CenteredLogo, Loading } from 'src/components';
 import ChangePlanForm from 'src/pages/billing/forms/ChangePlan';
 import Steps from './components/Steps';
 import { getPlans } from 'src/actions/account';
-import { getBillingCountries } from 'src/actions/billing';
+import { getBillingCountries, billingCreate } from 'src/actions/billing';
 
 export class OnboardingPlanPage extends Component {
   componentDidMount() {
     this.props.getPlans();
     this.props.getBillingCountries();
   }
+
+  onSubmit = (values) => {
+    const { billingCreate, showAlert, history } = this.props;
+    const uri = '/onboarding/sending-domain';
+
+    // no billing updates needed since they are still on free plan
+    if (values.planpicker.isFree) {
+      history.push(uri);
+      return;
+    }
+
+    return billingCreate(values)
+      .then(() => history.push(uri))
+      .then(() => showAlert({ type: 'success', message: 'Added plan' }))
+      .catch((err) => showAlert({ type: 'error', message: 'Selecting plan failed', details: err.message }));
+
+  };
 
   render() {
     const { loading } = this.props;
@@ -24,7 +42,7 @@ export class OnboardingPlanPage extends Component {
       <Fragment>
         <CenteredLogo />
         <Panel accent>
-          <ChangePlanForm onboarding={true} />
+          <ChangePlanForm onSubmit={this.onSubmit} onboarding={true} />
           <Steps />
         </Panel>
       </Fragment>
@@ -37,4 +55,4 @@ const mapStateToProps = (state) => ({
   loading: Boolean(state.account.loading || state.billing.plansLoading || state.billing.countriesLoading)
 });
 
-export default connect(mapStateToProps, { getPlans, getBillingCountries })(OnboardingPlanPage);
+export default connect(mapStateToProps, { billingCreate, showAlert, getPlans, getBillingCountries })(OnboardingPlanPage);
