+++ src/pages/onboarding/components/Steps.js
@@ -7,7 +7,6 @@ import styles from './Steps.module.scss';
 // TODO Finalize when whole register flow is complete
 const steps = [
   { path: 'plan', label: 'Confirm your Plan' },
-  { path: 'dedicated-ips', label: 'Add a Dedicated IP' },
   { path: 'sending-domain', label: 'Add a Domain' },
   { path: 'email', label: 'Send an Email' }
 ];
